+++ FurEverCarePlatform.Domain/Entities/BookingDetail.cs
@@ -7,9 +7,9 @@ public class BookingDetail : BaseEntity
     public Guid ServiceId { get; set; }
     public Guid PetId { get; set; }
     public required DateTime BookingTime { get; set; }
-    public decimal RealAmount { get; set; }
+    public float RealAmount { get; set; }
     public bool IsMeasured { get; set; }
-    public decimal RawAmount { get; set; }
+    public float RawAmount { get; set; }
     public float? PetWeight { get; set; }
     public string? Hair { get; set; }
     public Guid? AssignedUserId { get; set; }
